ExitController completed for generating bill and releasing parking slot

+++ parking-lot/src/main/java/com/practice/parking/services/BillCalculationStrategy.java
@@ -1,9 +1,7 @@
 package com.practice.parking.services;
 
-import com.practice.parking.models.Vehicle;
-
 import java.util.Date;
 
 public interface BillCalculationStrategy {
-    public int calculateAmount(Date entryTime, Date exitTime);
+    public long calculateAmount(Date entryTime, Date exitTime);
 }
